Release/prism v2.2 (#4)
* remove prism v1.4.0 files

* initial commit for prism v2.1

* disable .github workflow

* update .gitignore

* rename folder containing notebooks

* update docker-compose.yaml to new notebooks folder

* delete github workflow

* fix path for docker container

* update path to env files

* update path to env files and store credential record_id

* update path to env files and retrieve credential record_id

* fix playground and swagger paths

* add hashicorp vault and config for prism-agent

* add vault root token to run.sh

* generate v2.2 dereferenced openapi.yaml

* updated python client to Prism v2.2 (prism agent v1.5.0)

* update .gitignore

* add .env

* build: use correct image for hashicorp vault

* update PY and TS client to v1.6.0

* update script to deploy agents in WSL 2

* update notebooks to run with agent v1.6

* fix issue with schema author

* fix issue with python client

---------

+++ run_agents.sh
@@ -22,7 +22,11 @@ function getDockerHost() {
     local dockerHostAddress
     unset dockerHostAddress
     if [[ $(uname) == "Linux" ]] ; then
-      dockerHostAddress=$(docker run --rm --net=host eclipse/che-ip)
+      if [[ "$(< /proc/version)" == *@(Microsoft|WSL)* ]]; then
+        dockerHostAddress=host.docker.internal
+      else
+        dockerHostAddress=$(docker run --rm --net=host eclipse/che-ip)
+      fi
     else
       dockerHostAddress=host.docker.internal
     fi
